fix: Move `get_task_logs` to VFolderService
