+++ src/ai/backend/manager/services/vfolder/base.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+from typing import override
+
+from ai.backend.manager.actions.action import BaseAction
+
+
+@dataclass
+class VFolderAction(BaseAction):
+    @override
+    def entity_type(self):
+        return "vfolder"
